+++ History.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 0.8.0
+
+### Breaking changes
+* Removed old handler options which were deprecated in 0.6.0.
+
+### New features
+* Scrobble request which return error codes 11, 16 or 29 are automatically retried.
+
 ## 0.7.0
 
 ### Breaking changes